Rotate image: Fixed the rotating image issues.
git-svn-id: http://svn.contrexxlabs.com/contrexx/trunk@33120 5ce06e5d-2b28-0410-aa6a-b3cd9cf52d8a

+++ modules/Gallery/Controller/GalleryManager.class.php
@@ -3215,23 +3215,23 @@ $strFileNew = '';
         $objImage = new \ImageManager();
         $objImage->loadImage($strOrgPath.$strOldFilename);
         //Rotate the clockwise
-        $objImage->rotateImage(270);
-        $objImage->newImageType = \ImageManager::IMG_TYPE_PNG;
-        //To save the Rotated image
-        if($objImage->saveNewImage($strOrgPath.$strOldFilename, true)){
-            \Cx\Lib\FileSystem\FileSystem::delete_file($strThumbPath.$strOldFilename);
+        if ($objImage->rotateImage(270)) {
+            $objImage->newImageType = $objImage->orgImageType;
+            //To save the Rotated image
+            if ($objImage->saveNewImage($strOrgPath . $strOldFilename, true)) {
+                \Cx\Lib\FileSystem\FileSystem::delete_file($strThumbPath . $strOldFilename);
+            }
+            //Resize the Rotated image 
+            if ($objImage->resizeImageSave($strOrgPath, $strWebpath, $strOldFilename, $intX, $intY, $objResult->fields['quality'], $strThumbPath, $strThumbWebpath, $strOldFilename)) {
+                if ($objResult->fields['size_type'] == 'abs') {
+                    $objDatabase->Execute('    UPDATE     ' . DBPREFIX . 'module_gallery_pictures
+                                    SET     size_abs_h=' . $newInsertY . ',
+                                            size_abs_w=' . $newInsertX . '
+                                    WHERE     id=' . $intImageId);
+                }
+            }
         }
-        //Resize the Rotated image 
-        $objImage->resizeImageSave($strOrgPath,$strWebpath,$strOldFilename,$intX,$intY,$objResult->fields['quality'],
-                            $strThumbPath,$strThumbWebpath,$strOldFilename);
-        
-        if ($objResult->fields['size_type'] == 'abs') {
-            $objDatabase->Execute('    UPDATE     '.DBPREFIX.'module_gallery_pictures
-                                    SET     size_abs_h='.$newInsertY.',
-                                            size_abs_w='.$newInsertX.'
-                                    WHERE     id='.$intImageId);
-        }
-     }
+    }
 
 
     /**
